Perform VO2 kinetics fitting

Performs the fitting process for the VO2 kinetics analysis. At this point, the data should already have been cleaned (outliers removed) and processed (interpolated, time-aligned, ensembled-averaged, and bin-averaged).

Usage

perform_kinetics(
  .data_processed,
  intensity_domain = c("moderate", "heavy", "severe"),
  fit_level = 0.95,
  fit_phase_1_length,
  fit_baseline_length,
  fit_transition_length,
  verbose = TRUE,
  ...
)

Arguments

.data_processed

The data retrieved from process_data().

intensity_domain

The exercise-intensity domain that the test was performed. Either moderate, heavy, or severe.

fit_level

A numeric scalar between 0 and 1 giving the confidence level for the parameter estimates in the final VO2 kinetics fit. Default to 0.95.

fit_phase_1_length

The length of the phase I that you wish to exclude from the final exponential fit, in seconds. See VO2 kinetics section in ?vo2_kinetics for more details.

fit_baseline_length

The length the baseline to perform the final linear fit, in seconds. See VO2 kinetics section ?vo2_kinetics for more details.

fit_transition_length

The length of the transition to perform the final exponential fit, in seconds. See VO2 kinetics section ?vo2_kinetics for more details.

verbose

A boolean indicating whether messages should be printed in the console. Default to TRUE.

...

Additional arguments when fitting VO2 kinetics in the heavy- or severe-intensity domains. Arguments may be the following:

TODO

Value

a tibble containing one row and the nested columns:

data_fitted

The data containing the time and VO2 columns, as well as the fitted data and its residuals for each data point.

model

A nls object. The model used in the VO2 kinetics fitting.

model_summary

The tidied summary of the model.

model_residuals

The residuals of the model.

plot_model

The final plot of the fitted model.

plot_residuals

The residuals plot for the model diagnostics.

Details

See ?vo2_kinetics for details.