Process data for VO2 kinetics fitting

It removes the outliers detected through detect_outliers(), interpolates each transition, ensemble-averages all the transitions into one, performs a bin-average, and normalizes the time column (time zero will indicate the end of baseline and the start of the transition phase).

Usage

process_data(.data_outliers, protocol_baseline_length, fit_bin_average)

Arguments

.data_outliers

The data retrived from detect_outliers().

protocol_baseline_length

The length of the baseline (in seconds).

fit_bin_average

The bin average to be performed for the final fit.

Value

a tibble with the time-aligned, ensembled-averaged, and bin-averaged data.

Details

TODO