Work rate for a step-incremental test

This function produces the work rate throughout a step-incremental test given the protocol This will retrieve both the 'original' work rates, and also will perform a 'linearization' of the steps.

Usage

work_rate_step(
  .data,
  baseline_intensity,
  step_start,
  step_increase,
  step_length
)

Arguments

.data

The data with recognized protocol phases

baseline_intensity

The baseline intensity

step_start

In case the step test started in a different work rate than baseline

step_increase

The step in increase, in watts per step

step_length

The length, in seconds, of each step

Value

a tibble